moving viewmodels to its own package
